feat: Part4 exercises 4.15-4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,61 +1,62 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
-const { asyncHandler } = require('../utils/middleware');
+const { asyncHandler, userExtractor } = require('../utils/middleware');
+const responses = require('../utils/responses');
+const { checkTitleURLandUser, checkUserAndBlog } = require('../utils/blogService')
 
 blogsRouter.get('/', asyncHandler(async (req, res) => {
-    const blogs = await Blog.find({});
+    const blogs = await Blog
+        .find({})
+        .populate('user',{ name: 1, username: 1 });
     res.json(blogs);
 }));
 
-blogsRouter.post('/', asyncHandler(async (req, res) => {
-    const { author, title, url, likes } = req.body;
+blogsRouter.post('/', userExtractor, asyncHandler(async (req, res) => {
+    const check = await checkTitleURLandUser(req, res);
+    if (!check) return;
 
-    if (!title || !url) {
-        return res.status(400).json({
-            error: '400 Bad Request',
-            message: 'Title or URL is missing'
-        });
-    }
+    const { author, title, url, likes } = req.body;
 
     const blog = new Blog({
         author,
         title,
         url,
-        likes: likes ?? 0
+        likes: likes ?? 0,
+        user: req.user._id,
     });
 
     const savedBlog = await blog.save();
+    req.user.blogs = await req.user.blogs.concat(savedBlog._id);
+    await req.user.save();
+
     res.status(201).json(savedBlog);
 }));
 
-blogsRouter.delete('/:id', asyncHandler(async (req, res) => {
+blogsRouter.delete('/:id', userExtractor, asyncHandler(async (req, res) => {
+    const check = await checkUserAndBlog(req, res);
+    if (!check) return;
+
     const deletedBlog = await Blog.findByIdAndDelete(req.params.id);
 
-    if (!deletedBlog) {
-        return res.status(404).json({
-            error: '404 Not Found',
-            message: 'Blog does not exist'
-        });
-    }
+    req.user.blogs = req.user.blogs.filter(blogId =>
+        blogId.toString() !== req.params.id
+    );
+    await req.user.save();
 
-    res.status(204).json(deletedBlog);
+    return res.status(204).json(deletedBlog);
 }));
 
-blogsRouter.put('/:id', asyncHandler(async (req, res) => {
+blogsRouter.put('/:id', userExtractor, asyncHandler(async (req, res) => {
+    const check = await checkUserAndBlog(req, res);
+    if (!check) return;
+
     const updatedBlog = await Blog.findByIdAndUpdate(
         req.params.id,
         req.body,
         { new: true, runValidators: true }
     );
 
-    if (!updatedBlog) {
-        return res.status(404).json({
-            error: '404 Not Found',
-            message: 'Blog does not exist'
-        });
-    }
-
-    res.status(200).json(updatedBlog);
+    return res.status(200).json(updatedBlog);
 }));
 
 module.exports = blogsRouter;
